requirements.txt
llama-cpp-python
pydantic_settings
jsonpickle
confz